# docs/swagger.yaml
definitions:
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_whatsapp_CreateInstanceResponse:
    properties:
      data:
        $ref: '#/definitions/internal_whatsapp.CreateInstanceResponse'
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_whatsapp_GetInstancesResponse:
    properties:
      data:
        $ref: '#/definitions/internal_whatsapp.GetInstancesResponse'
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_whatsapp_connectResponse:
    properties:
      data:
        $ref: '#/definitions/internal_whatsapp.connectResponse'
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string:
    properties:
      data:
        $ref: '#/definitions/map_string_string'
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_whatsapp.CreateInstanceResponse:
    properties:
      instanceId:
        type: string
      pairingCode:
        type: string
      qrCode:
        type: string
    type: object
  internal_whatsapp.GetInstancesResponse:
    properties:
      instances:
        items:
          $ref: '#/definitions/internal_whatsapp.Instance'
        type: array
      userEmail:
        type: string
    type: object
  internal_whatsapp.Instance:
    properties:
      id:
        type: string
      instanceName:
        type: string
      phone:
        type: string
    required:
    - phone
    type: object
  internal_whatsapp.connectResponse:
    properties:
      code:
        type: string
      count:
        type: integer
      message:
        type: string
      pairingCode:
        type: string
      qrcode:
        properties:
          base64:
            type: string
          code:
            type: string
        type: object
      status:
        type: string
    type: object
  internal_whatsapp.createInstanceInput:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  map_string_string:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
paths:
...
  /whatsapp/instance:
    get:
      consumes:
      - application/json
      description: Busca todas as instâncias do WhatsApp para o usuário
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_whatsapp_GetInstancesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Busca todas as instâncias do WhatsApp
      tags:
      - whatsapp
    post:
      consumes:
      - application/json
      description: Cria uma nova instância do WhatsApp para o usuário
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_whatsapp.createInstanceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_whatsapp_CreateInstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Cria uma nova instância do WhatsApp
      tags:
      - whatsapp
  /whatsapp/instance/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a instância do usuário e permite criar/parear novamente.
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Deleta uma instância do WhatsApp
      tags:
      - whatsapp
  /whatsapp/instance/{id}/connect:
    post:
      consumes:
      - application/json
      description: Dispara a conexão usando number (telefone com DDI/DD).
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_whatsapp_connectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Conecta/pareia uma instância na Evolution
      tags:
      - whatsapp
  /whatsapp/instance/{id}/logout:
    delete:
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Desconecta/logout de uma instância na Evolution
      tags:
      - whatsapp
  /whatsapp/instance/{id}/restart:
    post:
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Reinicia uma instância na Evolution
      tags:
      - whatsapp
  /whatsapp/instance/{id}/status:
    get:
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Consulta status de conexão da instância
      tags:
      - whatsapp
swagger: "2.0"